feat/buy item from shop

+++ src/main/java/gr/nlamp/sfgame_backend/item/Item.java
@@ -45,5 +45,5 @@ public class Item implements Serializable {
     private Integer armor = 0;
 
     private BigInteger coinCost;
-    private BigInteger mushCost;
+    private Long mushCost;
 }
